fixed datePicker in movies update page fixed payment page to send actual price in mail changed type of payment to double

+++ MovieTime/client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/PurchaseCancellationPageController.java
@@ -20,7 +20,7 @@ public class PurchaseCancellationPageController {
 	private int waitingForMessageCounter;
 	private String orderNumber;
 	private Purchase foundPurchase;
-	private int refundAmount; // Needs to be changed to float...
+	private double refundAmount; // Needs to be changed to float...
 
 	@FXML
 	private TextField OrderNumberTextField;
